requirements.txt
# Core requirements
Flask==2.3.3
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0

# OpenAI
openai==1.3.0

# Development and debugging tools
pytest==7.4.0
flake8==6.1.0